Persisting form state after the startup is submitted.
Description:

This change updates the /startup/create page to persist form values after the form is submitted.

How:
Added a new field called data, in the default state of useActionState() hook. Once the user submits the form, the current form values are returned back to persist in the state.

Form elements like inputs, textarea is updated to consume defaultValue property, which points to the state.data